Preview

As a full run takes some time, it can be useful to run a quick preview to get meaning from the data and to get some guesses about the parameters of the full run. The only output of this step is preview.html, which visualizes the dataset and provides some diagnostics. Be careful, as this can get too large to be rendered for large datasets, so it's better to run on a subset of the data.

baysor preview <args> [options]

CLI parameters:

Baysor.CommandLine.preview — Function

Plot an html with the dataset preview.

Args

coordinates: CSV or Parquet file with coordinates of molecules and gene type

Options

-c, --config=<config.toml>: TOML file with a config. The function uses [data] and [plotting] sections.

-x, --x-column=<x>: Name of x column. Overrides the config value.

-y, --y-column=<y>: Name of y column. Overrides the config value.

-z, --z-column=<z>: Name of z column. Overrides the config value.

-g, --gene-column=<gene>: Name of gene column. Overrides the config value.

-m, --min-molecules-per-cell=<m>: Minimal number of molecules for a cell to be considered as real. It's an important parameter, as it's used to infer several other parameters. Overrides the config value.

-o, --output=<path>: Name of the output file or path to the output directory (default: "preview.html")