Segmentation-free analysis

Many analyses don't require segmentation, and can be run on local neighborhoods instead. In the paper, we call them Neighborhood Composition Vectors (NCVs). To obtain them from Baysor, you may run baysor segfree. This would output a loom file with NCVs of size k in (/matrix), storing ncv_color and confidence as column attributes (/col_attrs/).

baysor segfree <args> [options]

CLI parameters:

Baysor.CommandLine.segfree — Function

Extract Neighborhood Composition Vectors (NCVs) from a dataset

Args

coordinates: CSV or Parquet file with coordinates of molecules and gene type

Options

-k, --k-neighbors=<k>: Number of neighbors for segmentation-free pseudo-cells. It's suggested to set it to the expected minimal number of molecules per cell. (default: inferred from min-molecules-per-cell)

-c, --config=<config.toml>: TOML file with a config. The function only uses the [data] and [plotting] sections.

-x, --x-column=<x>: Name of x column. Overrides the config value.

-y, --y-column=<y>: Name of y column. Overrides the config value.

-z, --z-column=<z>: Name of z column. Overrides the config value.

-g, --gene-column=<gene>: Name of gene column. Overrides the config value.

-m, --min-molecules-per-cell=<m>: Minimal number of molecules for a cell to be considered as real. It's an important parameter, as it's used to infer several other parameters. Overrides the config value.

-o, --output=<path>: Name of the output file or path to the output directory (default: "ncvs.loom")